requirements.txt
ipywidgets==8.0.6
missingno==0.5.2
plotly==5.15.0
streamlit==1.23.1
streamlit-pandas-profiling==0.1.3
ydata-profiling==3.15.0
